Reorganized and added bootstrap
